Updated XUnit packages because current version was no longer working (https://github.com/Microsoft/vstest/issues/811). Removed HtmlAgilityPack (HAP) from the MalApi project. Added HAP to the integration and unit test projects. Moved update test logic to the integration test project. Doubled the scraping methods in the unit test project to keep the scraping method tests and updated the readme.txt in integration test project to include a warning/notification. Added test for non-existant MAL entry. Updating id that is not on your list does nothing but still returns "Updated" response.
