+++ MalApi.IntegrationTests/MalApi.IntegrationTests.csproj
@@ -11,10 +11,11 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\MalApi\MalApi.csproj" />
+    <PackageReference Include="HtmlAgilityPack" Version="1.7.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0" />
     <PackageReference Include="FluentAssertions" Version="4.19.2" />
-    <PackageReference Include="xunit" Version="2.3.0-beta3-build3705" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.3.0-beta3-build3705" />
+    <PackageReference Include="xunit" Version="2.3.1" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
   </ItemGroup>
   <ItemGroup Condition=" '$(TargetFramework)' == 'net452' ">
     <Reference Include="System" />
