+++ MalApi/MalApi.csproj
@@ -41,7 +41,6 @@
 
   <ItemGroup>
     <!-- Fixes breakage in .NET 4.6 apps caused by System.Net.Http 4.3.0. -->
-    <PackageReference Include="HtmlAgilityPack" Version="1.6.7" />
     <PackageReference Include="System.Net.Http" Version="4.3.1" />
     <PackageReference Include="Common.Logging" Version="3.4.0-Beta2" />
     <PackageReference Include="System.Xml.XmlSerializer" Version="4.3.0" />
